backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.2
annotated-types==0.5.0
appnope==0.1.3
argcomplete==1.10.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.8.2
blinker==1.6.2
Brotli==1.0.9
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.6
colorcet==3.0.1
comm==0.1.4
compressed-rtf==1.0.6
ConfigArgParse==1.7
cryptography==41.0.3
cssselect==1.2.0
dataclasses-json==0.5.14
debugpy==1.6.7
decorator==4.4.2
diskcache==5.6.1
docx2txt==0.8
ebcdic==1.1.1
exceptiongroup==1.1.2
executing==1.2.0
extract-msg==0.29.0
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Compress==1.13
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Sockets==0.2.1
Flask-SQLAlchemy==3.0.5
frozenlist==1.3.3
gevent==23.7.0
gunicorn==21.0.1
gevent-websocket==0.10.1
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.1
google-auth==2.22.0
googleapis-common-protos==1.60.0
goose3==3.1.17
greenlet==3.0.0a1
h11==0.14.0
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
IMAPClient==2.1.0
iniconfig==2.0.0
ipykernel==6.25.1
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.19.0
jieba==0.42.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.17.3
jupyter_client==8.3.0
jupyter_core==5.3.1
langchain==0.0.272
langdetect==1.0.9
langsmith==0.0.26
lxml==4.9.3
Mako==1.2.4
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.7
networkx==3.1
nltk==3.8.1
numexpr==2.8.5
numpy==1.24.3
olefile==0.46
openai==0.27.6
outcome==1.2.0
packaging==23.1
param==1.13.0
parso==0.8.3
pathspec==0.11.2
pdfminer.six==20191110
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
pluggy==1.2.0
proglog==0.1.10
prompt-toolkit==3.0.38
protobuf==4.24.1
psutil==5.9.5
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyahocorasick==2.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
pyct==0.5.0
pydantic==2.3.0
pydantic_core==2.6.3
Pygments==2.15.1
pyOpenSSL==23.2.0
pyrsistent==0.19.3
PySocks==1.7.1
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-pptx==0.6.21
PyYAML==6.0.1
pyzmq==25.1.0
regex==2023.8.8
requests==2.30.0
rich==13.3.5
rsa==4.9
scipy==1.10.1
six==1.12.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.4.1
SpeechRecognition==3.8.1
SQLAlchemy==2.0.19
stack-data==0.6.2
tenacity==8.2.3
tiktoken==0.4.0
tomli==2.0.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.2
trio-websocket==0.10.3
typing-inspect==0.9.0
typing_extensions==4.7.1
tzlocal==5.0.1
Unidecode==1.3.6
urllib3==1.26.16
urllib3-secure-extra==0.1.0
wcwidth==0.2.6
Werkzeug==2.3.7
wsproto==1.2.0
xlrd==1.2.0
XlsxWriter==3.1.2
yarl==1.9.2
zope.event==5.0
zope.interface==6.0
